refactor: hoist RuntimeName to @pnpm/types and rename system-version package
- Move RUNTIME_NAMES / RuntimeName / isRuntimeAlias into @pnpm/types so
  callers don't have to pull pkg-manifest.utils just for the alias check.
- Rename @pnpm/engine.runtime.system-node-version to
  @pnpm/engine.runtime.system-version now that it exports deno/bun
  helpers too. Directory + package name + all importers updated.
- Drop the duplicated "devEngines takes precedence" comment from the
  function body in favour of a docstring; convert the runtime-on-fail
  hint from a function to a const string.
