+++ .changeset/runtime-onfail-node-check.md
@@ -1,7 +1,8 @@
 ---
-"@pnpm/engine.runtime.system-node-version": minor
+"@pnpm/engine.runtime.system-version": minor
+"@pnpm/types": minor
 "@pnpm/config.reader": patch
 "pnpm": patch
 ---
 
-Validate `devEngines.runtime` and `engines.runtime` version ranges for `node`, `deno`, and `bun` when `onFail` is set to `error` or `warn`. Previously these settings only had an effect with `onFail: 'download'` — the `error` and `warn` modes silently did nothing [#11818](https://github.com/pnpm/pnpm/issues/11818).
+Validate `devEngines.runtime` and `engines.runtime` version ranges for `node`, `deno`, and `bun` when `onFail` is set to `error` or `warn`. Previously these settings only had an effect with `onFail: 'download'` — the `error` and `warn` modes silently did nothing [#11818](https://github.com/pnpm/pnpm/issues/11818). Violations now throw `ERR_PNPM_BAD_RUNTIME_VERSION`.
