+++ config/package-is-installable/package.json
@@ -34,7 +34,7 @@
   "dependencies": {
     "@pnpm/cli.meta": "workspace:*",
     "@pnpm/core-loggers": "workspace:*",
-    "@pnpm/engine.runtime.system-node-version": "workspace:*",
+    "@pnpm/engine.runtime.system-version": "workspace:*",
     "@pnpm/error": "workspace:*",
     "@pnpm/types": "workspace:*",
     "detect-libc": "catalog:",
