+++ core/types/src/package.ts
@@ -53,6 +53,14 @@ export interface DependenciesMeta {
   }
 }
 
+export const RUNTIME_NAMES = ['node', 'deno', 'bun'] as const
+
+export type RuntimeName = typeof RUNTIME_NAMES[number]
+
+export function isRuntimeAlias (alias: string): alias is RuntimeName {
+  return (RUNTIME_NAMES as readonly string[]).includes(alias)
+}
+
 export interface EngineDependency {
   name: string
   version?: string
