--- engine/runtime/system-node-version/README.md
@@ -1,15 +1,0 @@
-# @pnpm/engine.runtime.system-node-version
-
-> Detects the current system node version
-
-[![npm version](https://img.shields.io/npm/v/@pnpm/engine.runtime.system-node-version.svg)](https://npmx.dev/package/@pnpm/engine.runtime.system-node-version)
-
-## Installation
-
-```sh
-pnpm add @pnpm/engine.runtime.system-node-version
-```
-
-## License
-
-MIT
