+++ engine/runtime/system-version/tsconfig.json
@@ -11,6 +11,9 @@
   "references": [
     {
       "path": "../../../cli/meta"
+    },
+    {
+      "path": "../../../core/types"
     }
   ]
 }
