+++ exec/commands/tsconfig.json
@@ -61,7 +61,7 @@
       "path": "../../engine/runtime/commands"
     },
     {
-      "path": "../../engine/runtime/system-node-version"
+      "path": "../../engine/runtime/system-version"
     },
     {
       "path": "../../installing/client"
