+++ pnpm/package.json
@@ -100,7 +100,7 @@
     "@pnpm/deps.path": "workspace:*",
     "@pnpm/engine.pm.commands": "workspace:*",
     "@pnpm/engine.runtime.commands": "workspace:*",
-    "@pnpm/engine.runtime.system-node-version": "workspace:*",
+    "@pnpm/engine.runtime.system-version": "workspace:*",
     "@pnpm/error": "workspace:*",
     "@pnpm/exec.commands": "workspace:*",
     "@pnpm/hooks.pnpmfile": "workspace:*",
@@ -115,7 +115,6 @@
     "@pnpm/patching.commands": "workspace:*",
     "@pnpm/pkg-manifest.commands": "workspace:*",
     "@pnpm/pkg-manifest.reader": "workspace:*",
-    "@pnpm/pkg-manifest.utils": "workspace:*",
     "@pnpm/prepare": "workspace:*",
     "@pnpm/registry-mock": "catalog:",
     "@pnpm/registry-access.commands": "workspace:*",
